requirements/base.txt
channels[daphne]==4.0.0
channels-redis==4.1.0
Django==4.2.4
dj-database-url==2.1.0
dj-rest-auth==4.0.1
django-allauth==0.50.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-redis==5.3.0
django-unfold==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
Markdown==3.4.4
pandas==2.1.0
pymongo==4.5.0
psycopg[binary]==3.1.10
